\name{Vicus-package}
\alias{Vicus-package}
\docType{package}
\title{
\packageTitle{Vicus}
}
\description{
\packageDescription{Vicus}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Vicus}
\packageIndices{Vicus}
}
\author{
\packageAuthor{Vicus}

Maintainer: \packageMaintainer{Vicus}
}
\references{
Wang B, et al., (2017). Vicus: Exploiting local structures to improve network-based analysis of biological data. \emph{PLOS Computational Biology}. 13(10), e1005621

}

\keyword{ package }
\seealso{\code{\link{graphMatrix}},\code{\link{embedding}}}
\examples{
ls("package:Vicus")
}